refactor(#841): edge gateway concern folders (identity/gateway/protect/proxy/container)
Closes #841
Parent #829

No pilgrimage domain/. Typecheck green.
