+++ workers/edge/container/README.md
@@ -1,0 +1,3 @@
+# container
+
+Edge gateway concern folder (ticket #841). **No pilgrimage domain model.**
