+++ workers/edge/entry.test.ts
@@ -1,7 +1,7 @@
 import test from "node:test";
 import assert from "node:assert/strict";
 import { createWorkerApp, catalogOutbound } from "./app.ts";
-import { envWithCatalog, stubCtx } from "./entry-env.ts";
+import { envWithCatalog, stubCtx } from "./container/entry-env.ts";
 
 void test("GET /healthz reaches the container, not OpenNext", async () => {
   const app = createWorkerApp({});
