+++ workers/edge/gateway/README.md
@@ -1,0 +1,3 @@
+# gateway
+
+Edge gateway concern folder (ticket #841). **No pilgrimage domain model.**
