+++ workers/edge/identity/README.md
@@ -1,0 +1,3 @@
+# identity
+
+Edge gateway concern folder (ticket #841). **No pilgrimage domain model.**
