+++ workers/edge/protect/README.md
@@ -1,0 +1,3 @@
+# protect
+
+Edge gateway concern folder (ticket #841). **No pilgrimage domain model.**
