+++ workers/edge/proxy/README.md
@@ -1,0 +1,3 @@
+# proxy
+
+Edge gateway concern folder (ticket #841). **No pilgrimage domain model.**
